repgp: Adds `repgp' prefix to librepgp API functions
